separated reusable `AnimatedCircluarButton`
