+++ integration_test/purchases_test.dart
@@ -6,7 +6,6 @@ import 'package:lightmeter/data/models/ev_source_type.dart';
 import 'package:lightmeter/data/models/metering_screen_layout_config.dart';
 import 'package:lightmeter/data/shared_prefs_service.dart';
 import 'package:lightmeter/generated/l10n.dart';
-import 'package:lightmeter/screens/metering/components/bottom_controls/components/measure_button/widget_button_measure.dart';
 import 'package:lightmeter/screens/metering/components/shared/readings_container/components/equipment_profile_picker/widget_picker_equipment_profiles.dart';
 import 'package:lightmeter/screens/metering/components/shared/readings_container/components/extreme_exposure_pairs_container/widget_container_extreme_exposure_pairs.dart';
 import 'package:lightmeter/screens/metering/components/shared/readings_container/components/film_picker/widget_picker_film.dart';
@@ -21,6 +20,7 @@ import 'package:shared_preferences/shared_preferences.dart';
 
 import '../integration_test/utils/widget_tester_actions.dart';
 import 'mocks/iap_products_mock.dart';
+import 'utils/finder_actions.dart';
 
 @isTest
 void testPurchases(String description) {
@@ -84,7 +84,7 @@ void _expectProMeteringScreen({required bool enabled}) {
   expect(find.byType(NdValuePicker), findsOneWidget);
   expect(
     find.descendant(
-      of: find.byType(MeteringMeasureButton),
+      of: find.measureButton(),
       matching: find.byWidgetPredicate((widget) => widget is Text && widget.data!.contains('\u2081\u2080\u2080')),
     ),
     enabled ? findsOneWidget : findsNothing,
