+++ lib/screens/metering/components/bottom_controls/widget_bottom_controls.dart
@@ -2,8 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:lightmeter/data/models/ev_source_type.dart';
 import 'package:lightmeter/generated/l10n.dart';
 import 'package:lightmeter/providers/user_preferences_provider.dart';
-import 'package:lightmeter/screens/metering/components/bottom_controls/components/measure_button/widget_button_measure.dart';
+import 'package:lightmeter/screens/shared/animated_circular_button/widget_button_circular_animated.dart';
 import 'package:lightmeter/screens/shared/bottom_controls_bar/widget_bottom_controls_bar.dart';
+import 'package:lightmeter/utils/context_utils.dart';
 
 class MeteringBottomControls extends StatelessWidget {
   final double? ev;
@@ -39,11 +40,11 @@ class MeteringBottomControls extends StatelessWidget {
                   : S.of(context).tooltipUseLightSensor,
             )
           : null,
-      center: MeteringMeasureButton(
-        ev: ev,
-        ev100: ev100,
-        isMetering: isMetering,
-        onTap: onMeasure,
+      center: AnimatedCircluarButton(
+        progress: isMetering ? null : 1.0,
+        isPressed: isMetering,
+        onPressed: onMeasure,
+        child: ev != null ? _EvValueText(ev: ev!, ev100: ev100!) : null,
       ),
       right: IconButton(
         onPressed: onSettings,
@@ -51,5 +52,38 @@ class MeteringBottomControls extends StatelessWidget {
         tooltip: S.of(context).tooltipOpenSettings,
       ),
     );
+  }
+}
+
+class _EvValueText extends StatelessWidget {
+  static const String _subscript100 = '\u2081\u2080\u2080';
+  final double ev;
+  final double ev100;
+
+  const _EvValueText({
+    required this.ev,
+    required this.ev100,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+    return Text(
+      _text(context),
+      style: theme.textTheme.bodyMedium?.copyWith(color: theme.colorScheme.surface),
+      textAlign: TextAlign.center,
+    );
+  }
+
+  String _text(BuildContext context) {
+    final bool showEv100 = context.isPro && UserPreferencesProvider.showEv100Of(context);
+    final StringBuffer buffer = StringBuffer()
+      ..writeAll([
+        (showEv100 ? ev100 : ev).toStringAsFixed(1),
+        '\n',
+        S.of(context).ev,
+        if (showEv100) _subscript100,
+      ]);
+    return buffer.toString();
   }
 }
